requirements/base.txt
alabaster==0.7.12
amqp==2.6.1
appdirs==1.4.4
#argon2-cffi==19.2.0
asgiref==3.3.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
black==19.10b0
boto3==1.16.10
botocore==1.19.10
celery==4.4.2
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
Collectfast==2.1.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.5
django-allauth==0.41.0
django-anymail==7.1.0
django-braces==1.14.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-countries==6.1.3
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-crontab==0.7.1
django-datatables-too==0.1.3
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-extra-views==0.12.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-multi-form-view==2.0.1
django-redis==4.11.0
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-storages==1.9.1
django-stubs==1.5.0
django-timezone-field==4.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.16
entrypoints==0.3
factory-boy==2.12.0
Faker==4.14.0
filelock==3.0.12
flake8==3.7.9
flake8-isort==3.0.0
gunicorn==20.0.4
identify==1.5.6
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
ipdb==0.13.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
#Pillow==7.1.2
#pkg-resources==0.0.0
pluggy==0.13.1
pre-commit==2.3.0
prompt-toolkit==3.0.8
#psycopg2
#psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pylint-celery==0.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytest-sugar==0.9.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-http-client==3.3.1
python-slugify==4.0.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.4.7
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
sorl-thumbnail-serializer-field==0.2.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.48.0
termcolor==1.1.0
testfixtures==6.15.0
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
twilio==6.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.35
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
